requirements.txt
altair==6.0.0
annotated-types==0.7.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.7
choreographer==1.2.1
click==8.3.2
defusedxml==0.7.1
duckdb==1.5.2
fonttools==4.62.1
fpdf2==2.8.7
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kaleido==0.2.1
logistro==2.0.1
MarkupSafe==3.0.3
narwhals==2.19.0
numpy==2.4.4
orjson==3.11.8
packaging==26.1
pandas==3.0.2
patsy==1.0.2
pillow==12.2.0
plotly==6.7.0
pluggy==1.6.0
protobuf==7.34.1
pyarrow==23.0.1
pydantic==2.13.1
pydantic_core==2.46.1
pydeck==0.9.1
Pygments==2.20.0
pytest==9.0.3
pytest-html==4.2.0
pytest-metadata==3.1.1
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
simplejson==4.0.0
six==1.17.0
smmap==5.0.3
statsmodels==0.14.6
streamlit==1.56.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
watchdog==6.0.0
